requirements.txt
dearpygui==1.8.0
pyserial==3.5
